+++ src/lib/services/ContactsService.ts
@@ -680,8 +680,8 @@ class ContactsService extends GenericService {
      * @param {Object} dataToUpdate : 
      * { 
      * {string} number User phone number (as entered by user). Not mandatory if the PhoneNumber to update is a PhoneNumber linked to a system (pbx) Ordre de grandeur : 1..32 
-     * {string} type 	String Phone number type Valeurs autorisées : home, work, other
-     * {string} deviceType 	String Phone number device type Valeurs autorisées : landline, mobile, fax, other
+     * {string} type 	String Phone number type Possible values : home, work, other
+     * {string} deviceType 	String Phone number device type Possible values : landline, mobile, fax, other
      * {boolean} isVisibleByOthers optionnel 	Boolean Allow user to choose if the phone number is visible by other users or not. Note that administrators can see all the phone numbers, even if isVisibleByOthers is set to false. Note that phone numbers linked to a system (isFromSystem=true) are always visible, isVisibleByOthers can't be set to false for these numbers.
      * {string} shortNumber optionnel 	String [Only for update of PhoneNumbers linked to a system (pbx)] Short phone number (corresponds to the number monitored by PCG). Read only field, only used by server to find the related system PhoneNumber to update (couple shortNumber/systemId). Ordre de grandeur : 1..32
      * {string} systemId optionnel 	String [Only for update of PhoneNumbers linked to a system (pbx)] Unique identifier of the system in Rainbow database to which the system PhoneNumbers belong. Read only field, only used by server to find the related system PhoneNumber to update (couple shortNumber/systemId). Ordre de grandeur : 1..32
@@ -694,7 +694,7 @@ class ContactsService extends GenericService {
      * {string} nickName optionnel 	String User nickName Ordre de grandeur : 1..255
      * {string} title optionnel 	String User title (honorifics title, like Mr, Mrs, Sir, Lord, Lady, Dr, Prof,...) Ordre de grandeur : 1..40
      * {string} jobTitle optionnel 	String User job title Ordre de grandeur : 1..255
-     * {string} visibility optionnel 	String User visibility Define if the user can be searched by users being in other company and if the user can search users being in other companies. Visibility can be: same_than_company: The same visibility than the user's company's is applied to the user. When this user visibility is used, if the visibility of the company is changed the user's visibility will use this company new visibility. public: User can be searched by external users / can search external users. User can invite external users / can be invited by external users private: User can't be searched by external users / can search external users. User can invite external users / can be invited by external users closed: User can't be searched by external users / can't search external users. User can invite external users / can be invited by external users isolated: User can't be searched by external users / can't search external users. User can't invite external users / can't be invited by external users none: Default value reserved for guest. User can't be searched by any users (even within the same company) / can search external users. User can invite external users / can be invited by external users External users mean 'public user not being in user's company nor user's organisation nor a company visible by user's company. Valeur par défaut : same_than_company Valeurs autorisées : same_than_company, public, private, closed, isolated, none
+     * {string} visibility optionnel 	String User visibility Define if the user can be searched by users being in other company and if the user can search users being in other companies. Visibility can be: same_than_company: The same visibility than the user's company's is applied to the user. When this user visibility is used, if the visibility of the company is changed the user's visibility will use this company new visibility. public: User can be searched by external users / can search external users. User can invite external users / can be invited by external users private: User can't be searched by external users / can search external users. User can invite external users / can be invited by external users closed: User can't be searched by external users / can't search external users. User can invite external users / can be invited by external users isolated: User can't be searched by external users / can't search external users. User can't invite external users / can't be invited by external users none: Default value reserved for guest. User can't be searched by any users (even within the same company) / can search external users. User can invite external users / can be invited by external users External users mean 'public user not being in user's company nor user's organisation nor a company visible by user's company. Default value : same_than_company Possible values : same_than_company, public, private, closed, isolated, none
      * {boolean} isInitialized optionnel 	Boolean Is user initialized
      * {string} timezone optionnel 	String User timezone name Allowed values: one of the timezone names defined in IANA tz database Timezone name are composed as follow: Area/Location (ex: Europe/Paris, America/New_York,...)
      * {string} language optionnel 	String User language Language format is composed of locale using format ISO 639-1, with optionally the regional variation using ISO 3166‑1 alpha-2 (separated by hyphen). Locale part is in lowercase, regional part is in uppercase. Examples: en, en-US, fr, fr-FR, fr-CA, es-ES, es-MX, ... More information about the format can be found on this link. Ordre de grandeur : 2|5
@@ -703,7 +703,7 @@ class ContactsService extends GenericService {
      * {string} department optionnel 	String User department Ordre de grandeur : 1..255
      * {string} email 	String User email address Ordre de grandeur : 3..255
      * {string} country optionnel 	String Phone number country (ISO 3166-1 alpha3 format). country field is automatically computed using the following algorithm when creating/updating a phoneNumber entry: If number is provided and is in E164 format, country is computed from E164 number Else if country field is provided in the phoneNumber entry, this one is used Else user country field is used Note that in the case number field is set (but not in E164 format), associated numberE164 field is computed using phoneNumber'country field. So, number and country field must match so that numberE164 can be computed. Ordre de grandeur : 3
-     * {string} type 	String User email type Valeurs autorisées : home, work, other
+     * {string} type 	String User email type Possible values : home, work, other
      * {string} customData optionnel 	Object User's custom data. Object with free keys/values. It is up to the client to manage the user's customData (new customData provided overwrite the existing one). Restrictions on customData Object: max 20 keys, max key length: 64 characters, max value length: 4096 characters. User customData can only be created/updated by: the user himself `company_admin` or `organization_admin` of his company, `bp_admin` and `bp_finance` of his company, `superadmin`.
      * }
      * 
@@ -1116,6 +1116,322 @@ class ContactsService extends GenericService {
 
     //endregion Contacts NETWORK
 
+    //region Contacts Search
+
+    /**
+     * @public
+     * @method searchInAlldirectories
+     * @since 2.8.9
+     * @instance
+     * @category Contacts Search
+     * @description
+     * This API allows to search for resources matching given keywords. <br>
+     * Depending on the provided query parameters, search can be done: <br>
+     *   * on shortNumber
+     *   * on numberE164
+     * <br>
+     * <br>
+     * For both cases, systemId or pbxId must be provided, corresponding to the identifier of the system for which the search is requested. <br>
+     * <br>
+     * This API tries to find a resource in the directories: 
+     *   * PBX devices of the system for which the search is requested, if associated to a Rainbow user (PBX devices of all systems belonging to the system's group if applicable),
+     *   * phonebook of the system for which the search is requested (phonebooks of all systems belonging to the system's group if applicable),
+     *   * Office365 database associated to the company(ies) to which is(are) linked the system for which the search is requested,
+     *   * Business directory database associated to the company(ies) to which is(are) linked the system for which the search is requested.
+     *    <br>   
+     * If several entries match in several directories, the order defined in searchResultOrder setting of the system is applied. <br>
+     *    <br>
+     * @return {Promise<any>} An object of the result
+     *
+     * | Champ | Type | Description |
+     * | --- | --- | --- |
+     * | firstName | String | First name as it is present in one directory (either phonebbok,user or ActiveDirectory) |
+     * | lastName | String | Last name as it is present in one directory (either phonebbok,user or ActiveDirectory) |
+     * | id  | String | id of the user (if a user is found) |
+     * | jid_im | String | jid_im of the user (if a user is found) |
+     *
+     * @param {string} pbxId pbxId of the system for which the search is requested. One of systemId or pbxId is mandatory.
+     * @param {string} systemId identifier of the system for which the search is requested. One of systemId or pbxId is mandatory.
+     * @param {string} numberE164 Allows to filter users list on the numberE164 provided in this option.
+     * @param {string} shortnumber Allows to filter users list on the phone short number provided in this option.
+     * @param {string} format Allows to retrieve more or less phone book details in response. small: id, firstName, lastName, number. medium: id, firstName, lastName, number. full: id, firstName, lastName, number. Default value : small Possible values : small, medium, full.
+     * @param {number} limit Allow to specify the number of phone book entries to retrieve. Default value : 100
+     * @param {number} offset Allow to specify the position of first phone book entry to retrieve (first entry if not specified). Warning: if offset > total, no results are returned.
+     * @param {string} sortField Sort phone book list based on the given field. Default value : reverseDisplayName
+     * @param {number} sortOrder Specify order when sorting phone book list. Default value : 1. Possible values : -1, 1.
+     */
+    searchInAlldirectories (pbxId? : string, systemId? : string, numberE164? : string, shortnumber? : string, format : string = "small", limit : number = 100, offset? : number, sortField : string = "reverseDisplayName", sortOrder : number = 1) {
+        let that = this;
+
+        return new Promise((resolve, reject) => {
+
+            that._rest.searchInAlldirectories (pbxId, systemId, numberE164, shortnumber, format, limit, offset, sortField, sortOrder) .then(function (result) {
+                that._logger.log("info", LOG_ID + "(searchInAlldirectories) contact searched from server.");
+                that._logger.log("internal", LOG_ID + "(searchInAlldirectories) REST result : ", result);
+                return resolve(result);
+            }).catch(function (err) {
+                that._logger.log("error", LOG_ID + "(searchInAlldirectories) failed.");
+                that._logger.log("internalerror", LOG_ID + "(searchInAlldirectories) failed : ", err);
+                return reject(err);
+            });
+        });
+    }
+    
+    /**
+     * @public
+     * @method searchInPhonebook
+     * @since 2.8.9
+     * @instance
+     * @category Contacts Search
+     * @description
+     * This API allows to search for resources matching given keywords.The search is done on name and phone number. <br>
+     * Search can be: <br>
+     *   - on name: <br>
+     *      * keywords exact match (ex: 'John Doe' find 'John Doe')
+     *      * keywords partial match (ex: 'Jo Do' find 'John Doe')
+     *      * case insensitive (ex: 'john doe' find 'John Doe')
+     *      * accents insensitive (ex: 'herve mothe' find 'Hervé Mothé')
+     *      * on only firstname or lastname (ex: 'john' find 'John Doe')
+     *      * order firstname / lastname does not matter (ex: 'doe john' find 'John Doe')
+     *   - on number: <br>
+     *      * keywords exact match (ex: '0630594224' finds '0630594224' but NOT '+33630594224')
+     * <br>
+     * In case of user requesting the API, the search is done on user's system phonebook. <br>
+     * In case of PCG requesting the API, pbxId parameter is mandatory and the search is done on related system phonebook. <br>
+     *  <br>
+     * **Specific feature:** Sharing a system between several companies <br>
+     * Since 1.47.0 release, configuring companies sharing a multi-tenant system is possible. <br>
+     * An OXE can be multi-company. <br>
+     * A multi-tenant system, so called CENTREX, allows sharing a call-server between several entities. <br>
+     * Each company in this multi-tenant system has his own range of phone number. Each company has a company prefix named 'tenantCallNumber' in the companies data model <br>
+     *    <br>
+     *   - Company A - 8210xxxx (82103000 Alice, 82103001 Bob)
+     *   - Company B - 8211xxxx (82113001 Carol)
+     *    <br>
+     *   Carol can't search Alice by name because her phone number begins by a wrong company prefix.
+     *   Carol can't search Bob by number because her phone number begins by a wrong company prefix.".
+     *   In case of inconsistent configuration, an HTTP error 409210 is thrown.
+     *    <br>
+     * @return {Promise<any>} An object of the result
+     *
+     * | Champ | Type | Description |
+     * | --- | --- | --- |
+     * | id  | String | phone book id |
+     * | firstName | String | First name as it is present in the phone book |
+     * | lastName | String | Last name as it is present in the phone book |
+     * | number | String | Phone number as it is present in the phone book |
+     *
+     * @param {string} pbxId Mandatory if role is pcg.
+     * @param {string} name Allows to filter users list on the given keyword(s) provided in this option.
+     * @param {string} number Allows to filter users list on the phone number provided in this option.
+     * @param {string} format Allows to retrieve more or less phone book details in response. small: id, firstName, lastName, number. medium: id, firstName, lastName, number. full: id, firstName, lastName, number. Default value : small Possible values : small, medium, full.
+     * @param {number} limit Allow to specify the number of phone book entries to retrieve. Default value : 100
+     * @param {number} offset Allow to specify the position of first phone book entry to retrieve (first entry if not specified). Warning: if offset > total, no results are returned.
+     * @param {string} sortField Sort phone book list based on the given field. Default value : reverseDisplayName
+     * @param {number} sortOrder Specify order when sorting phone book list. Default value : 1. Possible values : -1, 1.
+     */
+    searchInPhonebook (pbxId : string, name : string, number : string, format : string, limit : number = 100, offset : number, sortField : string, sortOrder : number = 1) {
+        let that = this;
+
+        return new Promise((resolve, reject) => {
+
+            that._rest.searchInPhonebook (pbxId, name, number, format, limit, offset, sortField, sortOrder ).then(function (result) {
+                that._logger.log("info", LOG_ID + "(searchInPhonebook) contact searched from server.");
+                that._logger.log("internal", LOG_ID + "(searchInPhonebook) REST result : ", result);
+                return resolve(result);
+            }).catch(function (err) {
+                that._logger.log("error", LOG_ID + "(searchInPhonebook) failed.");
+                that._logger.log("internalerror", LOG_ID + "(searchInPhonebook) failed : ", err);
+                return reject(err);
+            });
+        });
+    }
+     
+    /**
+     * @public
+     * @method searchUserByPhonenumber
+     * @since 2.8.9
+     * @instance
+     * @category Contacts Search
+     * @description
+     * This API allows to search user being associated to the requested number. <br>
+     * The algorithm of this API is the following: <br>
+     *   * The API first search if the provided number matches one belonging to the pbx group of logged in user's pbx and being affected to a Rainbow user.
+     *   * Otherwise, the API search for users having the provided E164 number filled in their profile (only if setting isVisibleByOthers related to this number is not set to false). The API returns the result only if found user is in the same company or organisation than the logged in user's.
+     * If several numbers match, the first one found is returned. <br>
+     *    <br>
+     * @return {Promise<any>} An object of the result
+     * 
+     * | Champ | Type | Description |
+     * | --- | --- | --- |
+     * | id  | String | User unique identifier |
+     * | firstName | String | User first name |
+     * | lastName | String | User last name |
+     * | jid_im | String | User Jabber IM identifier |
+     * | companyId | String | User company unique identifier |
+     * | companyName | String | User company name |
+     * | organisationId | String | User organisation unique identifier |
+     * | lastAvatarUpdate | Date-Time | Date Date of last user avatar create/update, null if no avatar |
+     * | lastUpdateDate | Date-Time | Date of last user create/update |
+     * | guestMode | Boolean | Indicates a user embedded in a chat or conference room, as guest, with limited rights until he finalizes his registration. |
+     * | fileSharingCustomisation | String | Activate/Deactivate file sharing capability per user  <br>Define if the user can use the file sharing service.  <br>FileSharingCustomisation can be:<br><br>* `enabled`: The user can use the file sharing service.<br>* `disabled`: The user can't use the file sharing service. |
+     * | fileStorageCustomisation | String | Activate/Deactivate the capability for a user to access to Rainbow file storage.  <br>Define if a user has the right to upload/download/copy or share documents.  <br>fileStorageCustomisation can be:<br><br>* `enabled`: The user can manage and share files.<br>* `disabled`: The user can't manage and share files. |
+     * | useRoomCustomisation | String | Activate/Deactivate the capability for a user to use bubbles.  <br>Define if a user can create bubbles or participate in bubbles (chat and web conference).  <br>useRoomCustomisation can be:<br><br>* `enabled`: The user can use bubbles.<br>* `disabled`: The user can't use bubbles. |
+     * | phoneMeetingCustomisation | String | Activate/Deactivate the capability for a user to use phone meetings (PSTN conference).  <br>Define if a user has the right to join phone meetings.  <br>phoneMeetingCustomisation can be:<br><br>* `enabled`: The user can join phone meetings.<br>* `disabled`: The user can't join phone meetings. |
+     * | useChannelCustomisation | String | Activate/Deactivate the capability for a user to use a channel.  <br>Define if a user has the right to create channels or be a member of channels.  <br>useChannelCustomisation can be:<br><br>* `enabled`: The user can use some channels.<br>* `disabled`: The user can't use some channel. |
+     * | useScreenSharingCustomisation | String | Activate/Deactivate the capability for a user to share a screen.  <br>Define if a user has the right to share his screen.  <br>useScreenSharingCustomisation can be:<br><br>* `enabled`: Each user of the company can share his screen.<br>* `disabled`: No user of the company can share his screen. |
+     * | useWebRTCVideoCustomisation | String | Activate/Deactivate the capability for a user to switch to a Web RTC video conversation.  <br>Define if a user has the right to be joined via video and to use video (start a P2P video call, add video in a P2P call, add video in a web conference call).  <br>useWebRTCVideoCustomisation can be:<br><br>* `enabled`: The user can switch to a Web RTC video conversation.<br>* `disabled`: The user can't switch to a Web RTC video conversation. |
+     * | useWebRTCAudioCustomisation | String | Activate/Deactivate the capability for a user to switch to a Web RTC audio conversation.  <br>Define if a user has the right to be joined via audio (WebRTC) and to use Rainbow audio (WebRTC) (start a P2P audio call, start a web conference call).  <br>useWebRTCAudioCustomisation can be:<br><br>* `enabled`: The user can switch to a Web RTC audio conversation.<br>* `disabled`: The user can't switch to a Web RTC audio conversation. |
+     * | instantMessagesCustomisation | String | Activate/Deactivate the capability for a user to use instant messages.  <br>Define if a user has the right to use IM, then to start a chat (P2P ou group chat) or receive chat messages and chat notifications.  <br>instantMessagesCustomisation can be:<br><br>* `enabled`: The user can use instant messages.<br>* `disabled`: The user can't use instant messages. |
+     * | isTerminated | Boolean | Indicates if the user account has been deleted. |
+     *
+     * @param {number} number number to search. The number can be: <br>
+     *      - a system phone number being in the pbx group of logged in user's pbx <br>
+     *      - a phone number entered manually by a user in his profile and being in the same organisation than logged in user's (in that case, provided number must be in E164 format) <br>
+     *      
+     */
+    searchUserByPhonenumber(number : number ){
+        let that = this;
+
+        return new Promise((resolve, reject) => {
+            
+            that._rest.searchUserByPhonenumber(number).then(function (result) {
+                that._logger.log("info", LOG_ID + "(searchUserByPhonenumber) contact searched from server.");
+                that._logger.log("internal", LOG_ID + "(searchUserByPhonenumber) REST result : ", result);
+                return resolve(result);
+            }).catch(function (err) {
+                that._logger.log("error", LOG_ID + "(searchUserByPhonenumber) failed.");
+                that._logger.log("internalerror", LOG_ID + "(searchUserByPhonenumber) failed : ", err);
+                return reject(err);
+            });
+        });
+    }
+    
+    /**
+     * @public
+     * @method searchUsers
+     * @since 2.8.9
+     * @instance
+     * @category Contacts Search
+     * @description
+     *
+     * This API allows to search users.
+     * Two type of searches are available:
+     * * Search on displayName (query parameter `displayName`):
+     *      - The search is done on users' `firstName` and `lastName`, and search is done in all Rainbow public users and users being in companies visible by logged in user's company.
+     *      - If logged in user's has visibility `closed` or `isolated`, or `same_than_company` and logged in user's company has visibility `closed` or `isolated`, search is done only on users being in companies visible by logged in user's company.
+     *      - Search on display name can be:
+     *          * firstName and lastName exact match (ex: 'John Doe' find 'John Doe')
+     *          * partial match (ex: 'Jo Do' find 'John Doe')
+     *          * case insensitive (ex: 'john doe' find 'John Doe')
+     *          * accents insensitive (ex: 'herve mothe' find 'Hervé Mothé')
+     *          * on only firstname or lastname (ex: 'john' find 'John Doe')
+     *          * order firstname / lastname does not matter (ex: 'doe john' find 'John Doe')
+     *      - It is possible to specify on which company(ies) users are searched (using `companyId` query parameter)
+     *      - It is possible to exclude users from some company(ies) in the search (using `excludeCompanyId` query parameter)
+     * * Multi-criterion search (query parameter `search`):
+     *      - Multi criterion search is only available to users having feature `SEARCH_BY_TAGS` in their profiles,
+     *      - Multi criterion allows to search users based on their fields `firstName`, `lastName`, `jobTitle`, `department`, `companyName` and `tags`.
+     *          * Multi criterion search is limited to users belonging to logged in user's company or users being in a company that belongs to the same organization.
+     *          * For other users which does not belong to the same company or organisation (Rainbow public users and users being in companies visible by logged in user's company outside the organisation), the search is only done on users' `firstName` and `lastName`. If logged in user's has visibility `closed` or `isolated` (or `same_than_company` and logged in user's company has visibility `closed` or `isolated`), search on `firstName`/`lastName` is done only on users being in companies visible by logged in user's company (similar behavior than search with query parameter displayName).
+     *      - Provided search tags can be a single word or composed of several words separated by spaces.
+     *      - Only users matching all provided search tags in their fields `firstName`, `lastName`, `jobTitle`,`department`, `companyName` and/or `tags` (or `firstName` and/or `lastName` for users outside the logged in user company/organisation) will be returned in the results.
+     *      - Matching of the search tags is done from the start of the word, case is insensitive and special characters are ignored.
+     *      - Example, consider a user as follow:
+     *    <br>
+     *  { <br>
+     * firstName: 'John', <br>
+     * lastName: 'Doe', <br>
+     * companyName: 'Alcatel-Lucent International', <br>
+     * jobTitle: 'Sales Representative', <br>
+     * department: 'Sales', <br>
+     * tags: \['Healthcare', 'Hospitality'\] <br>
+     * } <br>
+     *    <br>
+     *  - This user can be found with the following search tags:
+     *      * exact match (ex: 'John Doe', 'John Sales Representative', 'John Healthcare', ...)
+     *      * partial match (ex: 'Jo Do', 'Jo Sales', 'Jo Health', 'Do Alcatel', ...)
+     *      * case insensitive (ex: 'john doe', 'john sales', 'john hospitality', 'doe alcatel', ...)
+     *      * on only one field (ex: 'doe', 'sales', 'healthcare')
+     *      * order does not matter (ex: 'doe john', 'sales alcatel', 'healthcare sales john', ...)
+     *  - It is possible to specify on which company(ies) users are searched (using companyId query parameter)
+     *  - It is possible to exclude users from some company(ies) in the search (using excludeCompanyId query parameter)
+     *    <br>
+     * One of `displayName` or `search` parameters must be provided to execute the search request.
+     *    <br>
+     * @return {Promise<any>} An object of the result
+     * 
+     * | Champ | Type | Description |
+     * | --- | --- | --- |
+     * | limit | Number | Number of requested items |
+     * | offset | Number | Requested position of the first item to retrieve |
+     * | total | Number | Total number of items |
+     * | data | Object\[\] | List of user Objects. |
+     * | id  | String | User unique identifier |
+     * | loginEmail | String | User email address (used for login)  <br>`loginEmail` field is only returned for users being in the same company than logged in user and not being in the default Rainbow company. |
+     * | firstName | String | User first name |
+     * | lastName | String | User last name |
+     * | jid_im | String | User Jabber IM identifier |
+     * | companyId | String | User company unique identifier |
+     * | companyName | String | User company name |
+     * | jobTitle optionnel | String | User job title.  <br>Only returned if search is requested using `search` parameter and found user is in the same company or organisation than logged in user. |
+     * | tags optionnel | String\[\] | Tags associated to the user by an administrator.  <br>Only returned if search is requested using `search` parameter and found user is in the same company or organisation than logged in user. |
+     * | lastAvatarUpdate | Date-Time | Date Date of last user avatar create/update, null if no avatar |
+     * | lastUpdateDate | Date-Time | Date of last user create/update |
+     * | guestMode | Boolean | Indicated a user embedded in a chat or conference room, as guest, with limited rights until he finalizes his registration. |
+     * | fileSharingCustomisation | String | Activate/Deactivate file sharing capability per user  <br>Define if the user can use the file sharing service.  <br>FileSharingCustomisation can be:<br><br>* `enabled`: The user can use the file sharing service.<br>* `disabled`: The user can't use the file sharing service. |
+     * | fileStorageCustomisation | String | Activate/Deactivate the capability for a user to access to Rainbow file storage.  <br>Define if a user has the right to upload/download/copy or share documents.  <br>fileStorageCustomisation can be:<br><br>* `enabled`: The user can manage and share files.<br>* `disabled`: The user can't manage and share files. |
+     * | useRoomCustomisation | String | Activate/Deactivate the capability for a user to use bubbles.  <br>Define if a user can create bubbles or participate in bubbles (chat and web conference).  <br>useRoomCustomisation can be:<br><br>* `enabled`: The user can use bubbles.<br>* `disabled`: The user can't use bubbles. |
+     * | phoneMeetingCustomisation | String | Activate/Deactivate the capability for a user to use phone meetings (PSTN conference).  <br>Define if a user has the right to join phone meetings.  <br>phoneMeetingCustomisation can be:<br><br>* `enabled`: The user can join phone meetings.<br>* `disabled`: The user can't join phone meetings. |
+     * | useChannelCustomisation | String | Activate/Deactivate the capability for a user to use a channel.  <br>Define if a user has the right to create channels or be a member of channels.  <br>useChannelCustomisation can be:<br><br>* `enabled`: The user can use some channels.<br>* `disabled`: The user can't use some channel. |
+     * | useScreenSharingCustomisation | String | Activate/Deactivate the capability for a user to share a screen.  <br>Define if a user has the right to share his screen.  <br>useScreenSharingCustomisation can be:<br><br>* `enabled`: Each user of the company can share his screen.<br>* `disabled`: No user of the company can share his screen. |
+     * | useWebRTCVideoCustomisation | String | Activate/Deactivate the capability for a user to switch to a Web RTC video conversation.  <br>Define if a user has the right to be joined via video and to use video (start a P2P video call, add video in a P2P call, add video in a web conference call).  <br>useWebRTCVideoCustomisation can be:<br><br>* `enabled`: The user can switch to a Web RTC video conversation.<br>* `disabled`: The user can't switch to a Web RTC video conversation. |
+     * | useWebRTCAudioCustomisation | String | Activate/Deactivate the capability for a user to switch to a Web RTC audio conversation.  <br>Define if a user has the right to be joined via audio (WebRTC) and to use Rainbow audio (WebRTC) (start a P2P audio call, start a web conference call).  <br>useWebRTCAudioCustomisation can be:<br><br>* `enabled`: The user can switch to a Web RTC audio conversation.<br>* `disabled`: The user can't switch to a Web RTC audio conversation. |
+     * | instantMessagesCustomisation | String | Activate/Deactivate the capability for a user to use instant messages.  <br>Define if a user has the right to use IM, then to start a chat (P2P ou group chat) or receive chat messages and chat notifications.  <br>instantMessagesCustomisation can be:<br><br>* `enabled`: The user can use instant messages.<br>* `disabled`: The user can't use instant messages. |
+     * | isTv | Boolean | True if it is a TV user |
+     * | isAlertNotificationEnabled | Boolean | True if user is to subscribed to Alert Offer |
+     * | phoneNumbers | Object\[\] | Array of user phone numbers objects.  <br>Phone number objects can:<br><br>* be created by user (information filled by user),<br>* come from association with a system (pbx) device (association is done by admin). |
+     * | phoneNumberId | String | Phone number unique id in phone-numbers directory collection. |
+     * | number optionnel | String | User phone number (as entered by user) |
+     * | numberE164 optionnel | String | User E.164 phone number, computed by server from `number` and `country` fields |
+     * | country | String | Phone number country (ISO 3166-1 alpha3 format)  <br>`country` field is automatically computed using the following algorithm when creating/updating a phoneNumber entry:<br><br>* If `number` is provided and is in E164 format, `country` is computed from E164 number<br>* Else if `country` field is provided in the phoneNumber entry, this one is used<br>* Else user `country` field is used |
+     * | isFromSystem optionnel | Boolean | Boolean indicating if phone is linked to a system (pbx). |
+     * | shortNumber optionnel | String | **\[Only for phone numbers linked to a system (pbx)\]**  <br>If phone is linked to a system (pbx), short phone number (corresponds to the number monitored by PCG).  <br>Only usable within the same PBX.  <br>Only PCG can set this field. |
+     * | internalNumber optionnel | String | **\[Only for phone numbers linked to a system (pbx)\]**  <br>If phone is linked to a system (pbx), internal phone number.  <br>Usable within a PBX group.  <br>Admins and users can modify this internalNumber field. |
+     * | systemId optionnel | String | **\[Only for phone numbers linked to a system (pbx)\]**  <br>If phone is linked to a system (pbx), unique identifier of that system in Rainbow database. |
+     * | pbxId optionnel | String | **\[Only for phone numbers linked to a system (pbx)\]**  <br>If phone is linked to a system (pbx), unique identifier of that pbx. |
+     * | type | String | Phone number type, one of `home`, `work`, `other`. |
+     * | deviceType | String | Phone number device type, one of `landline`, `mobile`, `fax`, `other`. |
+     * | isVisibleByOthers | Boolean | Allow user to choose if the phone number is visible by other users or not.  <br>Note that administrators can see all the phone numbers, even if `isVisibleByOthers` is set to false.  <br>Note that phone numbers linked to a system (`isFromSystem`=true) are always visible, `isVisibleByOthers` can't be set to false for these numbers. |
+     *
+     * @param {number} limit Allow to specify the number of users to retrieve. Default value : 20
+     * @param {string} displayName earch users on the given displayName. displayName and search parameters are exclusives, displayName parameter can only be set if search parameter is not provided.
+     * @param {string} search Search users belonging to the same company/organisation than logged in user on the given search tags on fields firstName, lastName, companyName, jobTitle, department,tags. Other public users/users in companies visible by logged in user's company are searched only on fields firstName and lastName (except if logged in user has visibility closed or isolated). displayName and search parameters are exclusives, search parameter can only be set if displayName parameter is not provided.
+     * @param {string} companyId Search users being in the requested company(ies). companyId and excludeCompanyId parameters are exclusives, companyId parameter can only be set if excludeCompanyId parameter is not provided.
+     * @param {string} excludeCompanyId Exclude users being in the requested company(ies) from the search results. companyId and excludeCompanyId parameters are exclusives, excludeCompanyId parameter can only be set if companyId parameter is not provided.
+     * @param {number} offset Allow to specify the position of first item to retrieve (first item if not specified). Warning: if offset > total, no results are returned.
+     * @param {string} sortField Sort items list based on the given field.
+     * @param {number} sortOrder Specify order when sorting items list. Default value : 1. Possible values : -1, 1
+     */    
+    searchUsers(limit : number = 20, displayName? : string, search? : string, companyId? : string, excludeCompanyId? : string, offset? : number, sortField? : string, sortOrder : number = 1){
+        let that = this;
+
+        return new Promise((resolve, reject) => {
+            
+            that._rest.searchUsers(limit, displayName, search, companyId, excludeCompanyId, offset, sortField, sortOrder).then(function (result) {
+                that._logger.log("info", LOG_ID + "(searchUsers) contact searched from server.");
+                that._logger.log("internal", LOG_ID + "(searchUsers) REST result : ", result);
+                return resolve(result);
+            }).catch(function (err) {
+                that._logger.log("error", LOG_ID + "(searchUsers) failed.");
+                that._logger.log("internalerror", LOG_ID + "(searchUsers) failed : ", err);
+                return reject(err);
+            });
+        });
+    }
+    
+    //endregion Contacts Search
+    
     //region Events
      
     /**